add sequenceOrder to netframework api
